Move config into the contribution
