+++ src/vs/workbench/contrib/terminalContrib/autoReplies/browser/terminal.autoReplies.contribution.ts
@@ -5,10 +5,11 @@
 
 import { Disposable } from '../../../../../base/common/lifecycle.js';
 import { IConfigurationService } from '../../../../../platform/configuration/common/configuration.js';
-import { TerminalSettingId, type ITerminalBackend } from '../../../../../platform/terminal/common/terminal.js';
+import { type ITerminalBackend } from '../../../../../platform/terminal/common/terminal.js';
 import { registerWorkbenchContribution2, WorkbenchPhase, type IWorkbenchContribution } from '../../../../common/contributions.js';
 import { ITerminalInstanceService } from '../../../terminal/browser/terminal.js';
-import { TERMINAL_CONFIG_SECTION, type ITerminalConfiguration } from '../../../terminal/common/terminal.js';
+import { TERMINAL_CONFIG_SECTION } from '../../../terminal/common/terminal.js';
+import { TerminalAutoRepliesSettingId, type ITerminalAutoRepliesConfiguration } from '../common/terminalAutoRepliesConfiguration.js';
 
 // #region Workbench contributions
 
@@ -29,7 +30,7 @@ export class TerminalAutoRepliesContribution extends Disposable implements IWork
 
 	private _installListenersOnBackend(backend: ITerminalBackend): void {
 		// Listen for config changes
-		const initialConfig = this._configurationService.getValue<ITerminalConfiguration>(TERMINAL_CONFIG_SECTION);
+		const initialConfig = this._configurationService.getValue<ITerminalAutoRepliesConfiguration>(TERMINAL_CONFIG_SECTION);
 		for (const match of Object.keys(initialConfig.autoReplies)) {
 			// Ensure the reply is value
 			const reply = initialConfig.autoReplies[match] as string | null;
@@ -38,12 +39,10 @@ export class TerminalAutoRepliesContribution extends Disposable implements IWork
 			}
 		}
 
-		// TODO: move to terminalconfigservice, add affectsConfiguration?
-		// TODO: Could simplify update to a single call
 		this._register(this._configurationService.onDidChangeConfiguration(async e => {
-			if (e.affectsConfiguration(TerminalSettingId.AutoReplies)) {
+			if (e.affectsConfiguration(TerminalAutoRepliesSettingId.AutoReplies)) {
 				backend.uninstallAllAutoReplies();
-				const config = this._configurationService.getValue<ITerminalConfiguration>(TERMINAL_CONFIG_SECTION);
+				const config = this._configurationService.getValue<ITerminalAutoRepliesConfiguration>(TERMINAL_CONFIG_SECTION);
 				for (const match of Object.keys(config.autoReplies)) {
 					// Ensure the reply is value
 					const reply = config.autoReplies[match] as string | null;
